+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* Added the clientID to the config.json file to make using the test server/live server interchangable.
+
+### Fixed
+* Added a null check on the phase checks.
 
  ## [1.0.1] 11/15/2018
 ### Added